Добавлены модели темизации для БД
