+++ packages/server/models/SiteTheme.ts
@@ -1,0 +1,31 @@
+import {
+  AllowNull,
+  AutoIncrement,
+  Column,
+  DataType,
+  Model,
+  PrimaryKey,
+  Table,
+  Unique
+} from 'sequelize-typescript'
+
+@Table({
+  timestamps: false,
+  paranoid: true,
+  tableName: 'site_theme',
+})
+export default class SiteTheme extends Model<SiteTheme> {
+  @AutoIncrement
+  @PrimaryKey
+  @Column(DataType.INTEGER)
+  override id: number
+
+  @AllowNull(false)
+  @Unique
+  @Column(DataType.STRING)
+  theme: string
+
+  @AllowNull(false)
+  @Column(DataType.STRING)
+  description: string
+}
