+++ packages/server/models/UserTheme.ts
@@ -1,0 +1,37 @@
+import {
+  AllowNull,
+  AutoIncrement,
+  Column,
+  DataType,
+  ForeignKey,
+  Index,
+  Model,
+  PrimaryKey,
+  Table
+} from 'sequelize-typescript'
+import SiteTheme from './SiteTheme'
+
+@Table({
+  timestamps: false,
+  paranoid: true,
+  tableName: 'user_theme',
+})
+export default class UserTheme extends Model<UserTheme> {
+  @AutoIncrement
+  @PrimaryKey
+  @Column(DataType.INTEGER)
+  override id: number
+
+  @Index
+  @AllowNull(false)
+  @Column(DataType.STRING)
+  theme: string
+
+  @Column(DataType.STRING)
+  device: string
+
+  @ForeignKey(() => SiteTheme)
+  @AllowNull(false)
+  @Column(DataType.INTEGER)
+  themeId: string
+}
