+++ packages/server/tsconfig.json
@@ -18,6 +18,9 @@
     "noUnusedLocals": true,
     "forceConsistentCasingInFileNames": true,
     "importsNotUsedAsValues": "error",
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "strictPropertyInitialization": false,
     "lib": ["es2019", "esnext.asynciterable", "dom"],
     "types": ["node", "jest"],
     "baseUrl": ".",
